Write latency is different in DDR2 and DDR4
In DDR2, the write latency is CAS-1. In DDR4, it is quite a bit more
complicated. Thus, I extracted the write latency as a separate parameter
and updated the implementation accordingly.

I have now manually reviewed the implementation of all state machine
transitions and checked that they they are in accordance with the DDR4
standard.