Partial lookahead component support
Optimizing for STP and HMoS requires using the speed-up and slow-down
relative to private mode, respectively. Added support for this in the
policy as well updated the UCP and ASM code such that they still work. 

Note: The lookahead algorithm implementation needs to be updated to work 
with both maximizing and minimizing the sum. Currently, it fails if a
metric that needs minimizing is selected.